refactor(Metrics): one Island for all metrics
We capture/record/sample these metrics based
on similar logic, so they can share an Island.
